requirements.txt
selenium~=4.35.0
allure-python-commons~=2.15.0
Faker~=37.6.0
pytest~=8.4.2